CONSUMPTION ENTRY MODULE: Added a comma feature in the qty fields.
This newly added feature includes, automatically format the entered value.

Previously the Qty entry field doesn't automatically format these kinds of values (1000, 10000, etc).

But these feature allows the value to be automatically format the value. So when the user enter a quantity or qty value 1000 or 10000, it will format this value into 1,000 and 10,000